+++ src/molecules/banner/Banner.vue
@@ -1,15 +1,13 @@
 <template src="./Banner.html" />
 
 <script>
-import ASource from '../source/Source.vue'
-import AImage from '../image/Image.vue'
+import AImage from '../../atoms/image/Image.vue'
 
 import ABanner from './Banner.js'
 
 export default {
   name: 'AlpacaBanner',
   components: {
-    ASource,
     AImage
   },
   mixins: [ABanner]