+++ src/molecules/cookie-message/CookieMessage.js
@@ -21,6 +21,13 @@ export default {
     closeAriaLabel: {
       type: String,
       default: null
+    },
+    /**
+     * Title for close icon
+     */
+    closeIconTitle: {
+      type: String,
+      default: null
     }
   },
   methods: {
@@ -29,7 +36,7 @@ export default {
 
       // after closing message box move focus on first focusable element on the page
       if (focusable.length) {
-        let firstFocusable = [...focusable].find(el => el.offsetParent !== null)
+        const firstFocusable = [...focusable].find(el => el.offsetParent !== null)
         firstFocusable.focus()
       }